Add Shared Diagnosis screen
