+++ app/src/main/java/org/covidwatch/android/di/AppModule.kt
@@ -36,6 +36,7 @@ import org.covidwatch.android.ui.exposures.ExposuresViewModel
 import org.covidwatch.android.ui.home.HomeViewModel
 import org.covidwatch.android.ui.menu.MenuViewModel
 import org.covidwatch.android.ui.onboarding.EnableExposureNotificationsViewModel
+import org.covidwatch.android.ui.reporting.SharedDiagnosisViewModel
 import org.covidwatch.android.ui.reporting.VerifyPositiveDiagnosisViewModel
 import org.covidwatch.android.ui.selectregion.SelectRegionViewModel
 import org.covidwatch.android.ui.settings.SettingsViewModel
@@ -291,6 +292,10 @@ val appModule = module {
 
     viewModel {
         MenuViewModel(exposureInformationRepository = get())
+    }
+
+    viewModel {
+        SharedDiagnosisViewModel(positiveDiagnosisRepository = get())
     }
 
     viewModel {
