+++ app/src/main/java/org/covidwatch/android/ui/reporting/BaseNotifyOthersFragment.kt
@@ -4,15 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.appcompat.app.AlertDialog
 import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.RecyclerView
 import org.covidwatch.android.R
-import org.covidwatch.android.databinding.DialogPastPositiveDiagnosesBinding
 import org.covidwatch.android.databinding.FragmentNotifyOthersBinding
-import org.covidwatch.android.extension.observe
 import org.covidwatch.android.extension.observeEvent
 import org.covidwatch.android.ui.BaseViewModelFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -22,8 +16,6 @@ open class BaseNotifyOthersFragment :
     BaseViewModelFragment<FragmentNotifyOthersBinding, NotifyOthersViewModel>() {
 
     override val viewModel: NotifyOthersViewModel by viewModel()
-
-    private val adapter = PositiveDiagnosisAdapter()
 
     override fun bind(
         inflater: LayoutInflater,
@@ -45,37 +37,10 @@ open class BaseNotifyOthersFragment :
             sharePositiveDiagnosisButton.setOnClickListener {
                 viewModel.sharePositiveDiagnosis()
             }
-
-            btnViewPastPositiveDiagnoses.setOnClickListener {
-                if (adapter.isEmpty()) {
-                    Toast.makeText(
-                        requireContext(),
-                        R.string.no_past_positive_diagnoses,
-                        Toast.LENGTH_SHORT
-                    ).show()
-                    return@setOnClickListener
-                }
-
-                val dialogView =
-                    DialogPastPositiveDiagnosesBinding.inflate(LayoutInflater.from(context))
-                dialogView.pastPositiveDiagnosesList.addItemDecoration(dividerItemDecoration())
-                dialogView.pastPositiveDiagnosesList.adapter = adapter
-                val dialog = AlertDialog
-                    .Builder(requireContext())
-                    .setView(dialogView.root)
-                    .create()
-                dialogView.closeButton.setOnClickListener { dialog.dismiss() }
-                dialog.show()
-            }
         }
 
         with(viewModel) {
-            observe(positiveDiagnosis) { adapter.setItems(it) }
             observeEvent(openVerificationScreen) { findNavController().navigate(R.id.verifyPositiveDiagnosisFragment) }
         }
-    }
-
-    private fun dividerItemDecoration(): RecyclerView.ItemDecoration {
-        return DividerItemDecoration(requireContext(), DividerItemDecoration.VERTICAL)
     }
 }
