--- app/src/main/java/org/covidwatch/android/ui/reporting/ThanksForReportingFragment.kt
@@ -1,30 +1,0 @@
-package org.covidwatch.android.ui.reporting
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
-import org.covidwatch.android.R
-import org.covidwatch.android.databinding.FragmentThanksForReportingBinding
-import org.covidwatch.android.extension.shareApp
-import org.covidwatch.android.ui.BaseFragment
-
-class ThanksForReportingFragment : BaseFragment<FragmentThanksForReportingBinding>() {
-
-    override fun bind(
-        inflater: LayoutInflater,
-        container: ViewGroup?
-    ): FragmentThanksForReportingBinding =
-        FragmentThanksForReportingBinding.inflate(inflater, container, false)
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.closeButton.setOnClickListener {
-            findNavController().popBackStack(R.id.homeFragment, false)
-        }
-        binding.shareAppButton.setOnClickListener {
-            context?.shareApp()
-        }
-    }
-}
