+++ app/src/main/res/drawable/next_step_bullet_point.xml
@@ -3,5 +3,5 @@
     <size
         android:width="8dp"
         android:height="8dp" />
-    <solid android:color="@color/plum" />
+    <solid android:color="@color/primaryText" />
 </shape>
