+++ app/src/main/res/layout/fragment_thanks_for_reporting.xml
@@ -21,16 +21,65 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="@dimen/padding_normal"
-            android:text="@string/step_3_notify_others" />
+            android:text="@string/diagnosis_shared_title" />
 
         <TextView
-            android:id="@+id/past_positive_diagnoses"
+            style="@style/TextView.ScreenSubtitle"
+            android:layout_marginHorizontal="@dimen/padding_normal"
+            android:text="@string/diagnosis_shared_subtitle"
+            android:textSize="16sp" />
+
+        <TextView
+            style="@style/TextView.ScreenSubtitle"
+            android:layout_marginHorizontal="@dimen/padding_normal"
+            android:text="@string/diagnosis_shared_next_steps_title"
+            android:textSize="16sp"
+            android:textStyle="bold" />
+
+        <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="@dimen/padding_normal"
-            android:layout_marginTop="@dimen/padding_large"
-            android:layout_weight="1"
+            android:layout_marginTop="@dimen/padding_normal"
+            android:drawableStart="@drawable/next_step_bullet_point"
+            android:drawablePadding="@dimen/padding_normal"
+            android:text="@string/diagnosis_shared_next_step_1"
             android:textSize="16sp" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/padding_normal"
+            android:layout_marginTop="@dimen/padding_normal"
+            android:drawableStart="@drawable/next_step_bullet_point"
+            android:drawablePadding="@dimen/padding_normal"
+            android:text="@string/diagnosis_shared_next_step_2"
+            android:textSize="16sp" />
+
+        <Button
+            android:id="@+id/share_app_button"
+            style="@style/Button.Action"
+            android:layout_marginHorizontal="@dimen/padding_normal"
+            android:layout_marginTop="@dimen/padding_normal"
+            android:text="@string/share_app_button" />
+
+        <Button
+            android:id="@+id/btn_view_past_positive_diagnoses"
+            style="@style/Widget.MaterialComponents.Button.OutlinedButton"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/padding_normal"
+            android:layout_marginTop="@dimen/padding_small"
+            android:paddingVertical="@dimen/padding_small"
+            android:text="@string/past_positive_diagnoses"
+            android:textAllCaps="false"
+            android:textColor="@color/colorPrimary"
+            android:textSize="@dimen/subtitle_font_size" />
+
+        <Space
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:layout_weight="1" />
 
         <ImageView
             android:id="@+id/imageView6"
@@ -39,18 +88,6 @@
             android:layout_marginTop="@dimen/padding_normal"
             android:contentDescription="@null"
             android:src="@drawable/reporting_art_3" />
-
-        <Space
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_weight="1" />
-
-        <Button
-            android:id="@+id/share_app_button"
-            style="@style/Button.Action"
-            android:layout_marginHorizontal="@dimen/padding_normal"
-            android:layout_marginTop="@dimen/padding_normal"
-            android:text="@string/share_app_button" />
 
         <include layout="@layout/powered_by" />
     </LinearLayout>
