+++ app/src/production/java/org/covidwatch/android/ui/reporting/NotifyOthers.kt
@@ -1,12 +1,4 @@
 package org.covidwatch.android.ui.reporting
 
-import org.covidwatch.android.data.positivediagnosis.PositiveDiagnosisRepository
-import org.covidwatch.android.exposurenotification.ExposureNotificationManager
-
 class NotifyOthersFragment : BaseNotifyOthersFragment()
-
-class NotifyOthersViewModel(
-    positiveDiagnosisRepository: PositiveDiagnosisRepository
-) : BaseNotifyOthersViewModel(
-    positiveDiagnosisRepository
-)
+class NotifyOthersViewModel : BaseNotifyOthersViewModel()
